requirements.txt
# Python Libraries 

flask==2.3.2
flask_cors==4.0.0
PyGithub==1.59.0
langchain==0.0.239
openai
